+++ RandomSkunk.Results/InvalidStateException.cs
@@ -16,39 +16,19 @@ public class InvalidStateException : Exception
     /// <summary>
     /// Initializes a new instance of the <see cref="InvalidStateException"/> class.
     /// </summary>
-    public InvalidStateException()
-    {
-    }
-
-    /// <summary>
-    /// Initializes a new instance of the <see cref="InvalidStateException"/> class.
-    /// </summary>
     /// <param name="message">The error message that explains the reason for the exception.</param>
-    public InvalidStateException(string message)
+    /// <param name="error">
+    /// The error that is the cause of the current exception, or a null reference if no error is
+    /// specified.
+    /// </param>
+    public InvalidStateException(string message, Error? error = null)
         : base(message)
     {
+        Error = error;
     }
 
     /// <summary>
-    /// Initializes a new instance of the <see cref="InvalidStateException"/> class.
+    /// Gets the error that is the cause of the current exception.
     /// </summary>
-    /// <param name="message">The error message that explains the reason for the exception.</param>
-    /// <param name="innerException">
-    /// The exception that is the cause of the current exception, or a null reference if no inner
-    /// exception is specified.
-    /// </param>
-    public InvalidStateException(string message, Exception innerException)
-        : base(message, innerException)
-    {
-    }
-
-    /// <summary>
-    /// Initializes a new instance of the <see cref="InvalidStateException"/> class with serialized data.
-    /// </summary>
-    /// <param name="info">The object that holds the serialized object data.</param>
-    /// <param name="context">The contextual information about the source or destination.</param>
-    protected InvalidStateException(SerializationInfo info, StreamingContext context)
-        : base(info, context)
-    {
-    }
+    public Error? Error { get; }
 }