+++ RandomSkunk.Results/Maybe{T}.cs
@@ -1,3 +1,5 @@
+using static RandomSkunk.Results.Error;
+
 namespace RandomSkunk.Results;
 
 /// <summary>
@@ -7,7 +9,7 @@ namespace RandomSkunk.Results;
 /// <remarks>
 /// Use <see cref="Create"/> to create instances of this type.
 /// </remarks>
-public struct Maybe<T> : IEquatable<Maybe<T>>
+public partial struct Maybe<T> : IEquatable<Maybe<T>>
 {
     /// <summary>
     /// The factory object used to create instances of <see cref="Maybe{T}"/>. This field is
@@ -75,11 +77,6 @@ public struct Maybe<T> : IEquatable<Maybe<T>>
     /// </returns>
     public bool IsFail => _type == MaybeType.Fail;
 
-    [NotNull]
-    internal T Value => _value!;
-
-    internal Error Error => _error ?? Error.DefaultError;
-
     /// <summary>
     /// Indicates whether the <paramref name="left"/> parameter is equal to the
     /// <paramref name="right"/> parameter.
@@ -122,6 +119,11 @@ public struct Maybe<T> : IEquatable<Maybe<T>>
         hashCode *= 31;
         return hashCode;
     }
+
+    internal Error Error() => _error ?? DefaultError;
+
+    [return: NotNull]
+    internal T Value() => _value!;
 
     private sealed class Factory : IMaybeFactory<T>
     {