+++ RandomSkunk.Results/Result.Match.cs
@@ -1,0 +1,194 @@
+namespace RandomSkunk.Results;
+
+/// <content>
+/// Defines the <c>Match</c> methods.
+/// </content>
+public partial struct Result
+{
+    /// <summary>
+    /// Evaluates either the <paramref name="success"/> or <paramref name="fail"/>
+    /// function depending on whether the result type is <c>Success</c> or <c>Fail</c>.
+    /// </summary>
+    /// <typeparam name="T">The return type of the match method.</typeparam>
+    /// <param name="success">
+    /// The function to evaluate if the result type is <c>Success</c>.
+    /// </param>
+    /// <param name="fail">
+    /// The function to evaluate if the result type is <c>Fail</c>. The error message and error
+    /// code of the <c>Fail</c> result are passed to this function.
+    /// </param>
+    /// <returns>The result of the matching function evaluation.</returns>
+    /// <exception cref="ArgumentNullException">
+    /// If <paramref name="success"/> is <see langword="null"/> or if <paramref name="fail"/> is
+    /// <see langword="null"/>.
+    /// </exception>
+    public T Match<T>(
+        Func<T> success,
+        Func<Error, T> fail)
+    {
+        if (success is null) throw new ArgumentNullException(nameof(success));
+        if (fail is null) throw new ArgumentNullException(nameof(fail));
+
+        return IsSuccess
+            ? success()
+            : fail(Error());
+    }
+
+    /// <summary>
+    /// Evaluates either the <paramref name="success"/> or <paramref name="fail"/>
+    /// function depending on whether the result type is <c>Success</c> or <c>Fail</c>.
+    /// </summary>
+    /// <param name="success">
+    /// The function to evaluate if the result type is <c>Success</c>.
+    /// </param>
+    /// <param name="fail">
+    /// The function to evaluate if the result type is <c>Fail</c>. The error message and error
+    /// code of the <c>Fail</c> result are passed to this function.
+    /// </param>
+    /// <exception cref="ArgumentNullException">
+    /// If <paramref name="success"/> is <see langword="null"/> or if <paramref name="fail"/> is
+    /// <see langword="null"/>.
+    /// </exception>
+    public void Match(
+        Action success,
+        Action<Error> fail)
+    {
+        if (success is null) throw new ArgumentNullException(nameof(success));
+        if (fail is null) throw new ArgumentNullException(nameof(fail));
+
+        if (IsSuccess)
+            success();
+        else
+            fail(Error());
+    }
+
+    /// <summary>
+    /// Evaluates either the <paramref name="success"/> or <paramref name="fail"/>
+    /// function depending on whether the result type is <c>Success</c> or <c>Fail</c>.
+    /// </summary>
+    /// <typeparam name="T">The return type of the match method.</typeparam>
+    /// <param name="success">
+    /// The function to evaluate if the result type is <c>Success</c>.
+    /// </param>
+    /// <param name="fail">
+    /// The function to evaluate if the result type is <c>Fail</c>. The error message and error
+    /// code of the <c>Fail</c> result are passed to this function.
+    /// </param>
+    /// <param name="cancellationToken">
+    /// The token to monitor for cancellation requests. The default value is
+    /// <see cref="CancellationToken.None"/>.
+    /// </param>
+    /// <returns>
+    /// A task that represents the asynchronous match operation, which wraps the result of the
+    /// matching function evaluation.
+    /// </returns>
+    /// <exception cref="ArgumentNullException">
+    /// If <paramref name="success"/> is <see langword="null"/> or if <paramref name="fail"/> is
+    /// <see langword="null"/>.
+    /// </exception>
+    public Task<T> MatchAsync<T>(
+        Func<CancellationToken, Task<T>> success,
+        Func<Error, CancellationToken, Task<T>> fail,
+        CancellationToken cancellationToken = default)
+    {
+        if (success is null) throw new ArgumentNullException(nameof(success));
+        if (fail is null) throw new ArgumentNullException(nameof(fail));
+
+        return IsSuccess
+            ? success(cancellationToken)
+            : fail(Error(), cancellationToken);
+    }
+
+    /// <summary>
+    /// Evaluates either the <paramref name="success"/> or <paramref name="fail"/>
+    /// function depending on whether the result type is <c>Success</c> or <c>Fail</c>.
+    /// </summary>
+    /// <typeparam name="T">The return type of the match method.</typeparam>
+    /// <param name="success">
+    /// The function to evaluate if the result type is <c>Success</c>.
+    /// </param>
+    /// <param name="fail">
+    /// The function to evaluate if the result type is <c>Fail</c>. The error message and error
+    /// code of the <c>Fail</c> result are passed to this function.
+    /// </param>
+    /// <returns>
+    /// A task that represents the asynchronous match operation, which wraps the result of the
+    /// matching function evaluation.
+    /// </returns>
+    /// <exception cref="ArgumentNullException">
+    /// If <paramref name="success"/> is <see langword="null"/> or if <paramref name="fail"/> is
+    /// <see langword="null"/>.
+    /// </exception>
+    public Task<T> MatchAsync<T>(
+        Func<Task<T>> success,
+        Func<Error, Task<T>> fail)
+    {
+        if (success is null) throw new ArgumentNullException(nameof(success));
+        if (fail is null) throw new ArgumentNullException(nameof(fail));
+
+        return MatchAsync(
+            _ => success(),
+            (error, _) => fail(error));
+    }
+
+    /// <summary>
+    /// Evaluates either the <paramref name="success"/> or <paramref name="fail"/>
+    /// function depending on whether the result type is <c>Success</c> or <c>Fail</c>.
+    /// </summary>
+    /// <param name="success">
+    /// The function to evaluate if the result type is <c>Success</c>.
+    /// </param>
+    /// <param name="fail">
+    /// The function to evaluate if the result type is <c>Fail</c>. The error message and error
+    /// code of the <c>Fail</c> result are passed to this function.
+    /// </param>
+    /// <param name="cancellationToken">
+    /// The token to monitor for cancellation requests. The default value is
+    /// <see cref="CancellationToken.None"/>.
+    /// </param>
+    /// <returns>A task representing the asynchronous match operation.</returns>
+    /// <exception cref="ArgumentNullException">
+    /// If <paramref name="success"/> is <see langword="null"/> or if <paramref name="fail"/> is
+    /// <see langword="null"/>.
+    /// </exception>
+    public Task MatchAsync(
+        Func<CancellationToken, Task> success,
+        Func<Error, CancellationToken, Task> fail,
+        CancellationToken cancellationToken = default)
+    {
+        if (success is null) throw new ArgumentNullException(nameof(success));
+        if (fail is null) throw new ArgumentNullException(nameof(fail));
+
+        return IsSuccess
+            ? success(cancellationToken)
+            : fail(Error(), cancellationToken);
+    }
+
+    /// <summary>
+    /// Evaluates either the <paramref name="success"/> or <paramref name="fail"/>
+    /// function depending on whether the result type is <c>Success</c> or <c>Fail</c>.
+    /// </summary>
+    /// <param name="success">
+    /// The function to evaluate if the result type is <c>Success</c>.
+    /// </param>
+    /// <param name="fail">
+    /// The function to evaluate if the result type is <c>Fail</c>. The error message and error
+    /// code of the <c>Fail</c> result are passed to this function.
+    /// </param>
+    /// <returns>A task representing the asynchronous match operation.</returns>
+    /// <exception cref="ArgumentNullException">
+    /// If <paramref name="success"/> is <see langword="null"/> or if <paramref name="fail"/> is
+    /// <see langword="null"/>.
+    /// </exception>
+    public Task MatchAsync(
+        Func<Task> success,
+        Func<Error, Task> fail)
+    {
+        if (success is null) throw new ArgumentNullException(nameof(success));
+        if (fail is null) throw new ArgumentNullException(nameof(fail));
+
+        return MatchAsync(
+            _ => success(),
+            (error, _) => fail(error));
+    }
+}