+++ RandomSkunk.Results/ResultExtensions.Equals.cs
@@ -22,7 +22,7 @@ public static partial class ResultExtensions
     {
         comparer ??= EqualityComparer<T>.Default;
 
-        return source.IsSuccess && comparer.Equals(source.Value, otherValue);
+        return source.IsSuccess && comparer.Equals(source.Value(), otherValue);
     }
 
     /// <summary>
@@ -46,7 +46,7 @@ public static partial class ResultExtensions
     {
         if (isSuccessValue is null) throw new ArgumentNullException(nameof(isSuccessValue));
 
-        return source.IsSuccess && isSuccessValue(source.Value);
+        return source.IsSuccess && isSuccessValue(source.Value());
     }
 
     /// <summary>
@@ -67,7 +67,7 @@ public static partial class ResultExtensions
     {
         comparer ??= EqualityComparer<T>.Default;
 
-        return source.IsSome && comparer.Equals(source.Value, otherValue);
+        return source.IsSome && comparer.Equals(source.Value(), otherValue);
     }
 
     /// <summary>
@@ -91,6 +91,6 @@ public static partial class ResultExtensions
     {
         if (isSomeValue is null) throw new ArgumentNullException(nameof(isSomeValue));
 
-        return source.IsSome && isSomeValue(source.Value);
+        return source.IsSome && isSomeValue(source.Value());
     }
 }