+++ RandomSkunk.Results/Result{T}.Match.cs
@@ -1,0 +1,200 @@
+namespace RandomSkunk.Results;
+
+/// <content>
+/// Defines the <c>Match</c> methods.
+/// </content>
+public partial struct Result<T>
+{
+    /// <summary>
+    /// Evaluates either the <paramref name="success"/> or <paramref name="fail"/>
+    /// function depending on whether the result type is <c>Success</c> or <c>Fail</c>.
+    /// </summary>
+    /// <typeparam name="TReturn">The return type of the match method.</typeparam>
+    /// <param name="success">
+    /// The function to evaluate if the result type is <c>Success</c>. The value of the <c>Success</c> result
+    /// is passed to this function.
+    /// </param>
+    /// <param name="fail">
+    /// The function to evaluate if the result type is <c>Fail</c>. The error message and error code
+    /// of the <c>Fail</c> result are passed to this function.
+    /// </param>
+    /// <returns>The result of the matching function evaluation.</returns>
+    /// <exception cref="ArgumentNullException">
+    /// If <paramref name="success"/> is <see langword="null"/> or if <paramref name="fail"/> is
+    /// <see langword="null"/>.
+    /// </exception>
+    public TReturn Match<TReturn>(
+        Func<T, TReturn> success,
+        Func<Error, TReturn> fail)
+    {
+        if (success is null) throw new ArgumentNullException(nameof(success));
+        if (fail is null) throw new ArgumentNullException(nameof(fail));
+
+        return IsSuccess
+            ? success(Value())
+            : fail(Error());
+    }
+
+    /// <summary>
+    /// Evaluates either the <paramref name="success"/> or <paramref name="fail"/>
+    /// function depending on whether the result type is <c>Success</c> or <c>Fail</c>.
+    /// </summary>
+    /// <param name="success">
+    /// The function to evaluate if the result type is <c>Success</c>. The value of the <c>Success</c> result
+    /// is passed to this function.
+    /// </param>
+    /// <param name="fail">
+    /// The function to evaluate if the result type is <c>Fail</c>. The error message and error code
+    /// of the <c>Fail</c> result are passed to this function.
+    /// </param>
+    /// <exception cref="ArgumentNullException">
+    /// If <paramref name="success"/> is <see langword="null"/> or if <paramref name="fail"/> is
+    /// <see langword="null"/>.
+    /// </exception>
+    public void Match(
+        Action<T> success,
+        Action<Error> fail)
+    {
+        if (success is null) throw new ArgumentNullException(nameof(success));
+        if (fail is null) throw new ArgumentNullException(nameof(fail));
+
+        if (IsSuccess)
+            success(Value());
+        else
+            fail(Error());
+    }
+
+    /// <summary>
+    /// Evaluates either the <paramref name="success"/> or <paramref name="fail"/>
+    /// function depending on whether the result type is <c>Success</c> or <c>Fail</c>.
+    /// </summary>
+    /// <typeparam name="TReturn">The return type of the match method.</typeparam>
+    /// <param name="success">
+    /// The function to evaluate if the result type is <c>Success</c>. The value of the <c>Success</c> result
+    /// is passed to this function.
+    /// </param>
+    /// <param name="fail">
+    /// The function to evaluate if the result type is <c>Fail</c>. The error message and error code
+    /// of the <c>Fail</c> result are passed to this function.
+    /// </param>
+    /// <param name="cancellationToken">
+    /// The token to monitor for cancellation requests. The default value is
+    /// <see cref="CancellationToken.None"/>.
+    /// </param>
+    /// <returns>
+    /// A task that represents the asynchronous match operation, which wraps the result of the
+    /// matching function evaluation.
+    /// </returns>
+    /// <exception cref="ArgumentNullException">
+    /// If <paramref name="success"/> is <see langword="null"/> or if <paramref name="fail"/> is
+    /// <see langword="null"/>.
+    /// </exception>
+    public Task<TReturn> MatchAsync<TReturn>(
+        Func<T, CancellationToken, Task<TReturn>> success,
+        Func<Error, CancellationToken, Task<TReturn>> fail,
+        CancellationToken cancellationToken = default)
+    {
+        if (success is null) throw new ArgumentNullException(nameof(success));
+        if (fail is null) throw new ArgumentNullException(nameof(fail));
+
+        return IsSuccess
+            ? success(Value(), cancellationToken)
+            : fail(Error(), cancellationToken);
+    }
+
+    /// <summary>
+    /// Evaluates either the <paramref name="success"/> or <paramref name="fail"/>
+    /// function depending on whether the result type is <c>Success</c> or <c>Fail</c>.
+    /// </summary>
+    /// <typeparam name="TReturn">The return type of the match method.</typeparam>
+    /// <param name="success">
+    /// The function to evaluate if the result type is <c>Success</c>. The value of the <c>Success</c> result
+    /// is passed to this function.
+    /// </param>
+    /// <param name="fail">
+    /// The function to evaluate if the result type is <c>Fail</c>. The error message and error code
+    /// of the <c>Fail</c> result are passed to this function.
+    /// </param>
+    /// <returns>
+    /// A task that represents the asynchronous match operation, which wraps the result of the
+    /// matching function evaluation.
+    /// </returns>
+    /// <exception cref="ArgumentNullException">
+    /// If <paramref name="success"/> is <see langword="null"/> or if <paramref name="fail"/> is
+    /// <see langword="null"/>.
+    /// </exception>
+    public Task<TReturn> MatchAsync<TReturn>(
+        Func<T, Task<TReturn>> success,
+        Func<Error, Task<TReturn>> fail)
+    {
+        if (success is null) throw new ArgumentNullException(nameof(success));
+        if (fail is null) throw new ArgumentNullException(nameof(fail));
+
+        return MatchAsync(
+            (value, _) => success(value),
+            (error, _) => fail(error));
+    }
+
+    /// <summary>
+    /// Evaluates either the <paramref name="success"/> or <paramref name="fail"/>
+    /// function depending on whether the result type is <c>Success</c> or <c>Fail</c>.
+    /// </summary>
+    /// <param name="success">
+    /// The function to evaluate if the result type is <c>Success</c>. The value of the <c>Success</c> result
+    /// is passed to this function.
+    /// </param>
+    /// <param name="fail">
+    /// The function to evaluate if the result type is <c>Fail</c>. The error message and error code
+    /// of the <c>Fail</c> result are passed to this function.
+    /// </param>
+    /// <param name="cancellationToken">
+    /// The token to monitor for cancellation requests. The default value is
+    /// <see cref="CancellationToken.None"/>.
+    /// </param>
+    /// <returns>A task representing the asynchronous match operation.</returns>
+    /// <exception cref="ArgumentNullException">
+    /// If <paramref name="success"/> is <see langword="null"/> or if <paramref name="fail"/> is
+    /// <see langword="null"/>.
+    /// </exception>
+    public Task MatchAsync(
+        Func<T, CancellationToken, Task> success,
+        Func<Error, CancellationToken, Task> fail,
+        CancellationToken cancellationToken = default)
+    {
+        if (success is null) throw new ArgumentNullException(nameof(success));
+        if (fail is null) throw new ArgumentNullException(nameof(fail));
+
+        return IsSuccess
+            ? success(Value(), cancellationToken)
+            : fail(Error(), cancellationToken);
+    }
+
+    /// <summary>
+    /// Evaluates either the <paramref name="success"/> or <paramref name="fail"/>
+    /// function depending on whether the result type is <c>Success</c> or <c>Fail</c>.
+    /// </summary>
+    /// <param name="success">
+    /// The function to evaluate if the result type is <c>Success</c>. The value of the <c>Success</c> result
+    /// is passed to this function.
+    /// </param>
+    /// <param name="fail">
+    /// The function to evaluate if the result type is <c>Fail</c>. The error message and error code
+    /// of the <c>Fail</c> result are passed to this function.
+    /// </param>
+    /// <returns>A task representing the asynchronous match operation.</returns>
+    /// <exception cref="ArgumentNullException">
+    /// If <paramref name="success"/> is <see langword="null"/> or if <paramref name="fail"/> is
+    /// <see langword="null"/>.
+    /// </exception>
+    public Task MatchAsync(
+        Func<T, Task> success,
+        Func<Error, Task> fail)
+    {
+        if (success is null) throw new ArgumentNullException(nameof(success));
+        if (fail is null) throw new ArgumentNullException(nameof(fail));
+
+        return MatchAsync(
+            (value, _) => success(value),
+            (error, _) => fail(error));
+    }
+}